Clean up test code
Move it to Storybook or similar later
