+++ client/src/components/dev/main-dev.tsx
@@ -1,0 +1,9 @@
+import Sockets from "./sockets";
+
+export default function MainDev() {
+  return process.env.NODE_ENV === 'development' && (
+    <>
+      <Sockets />
+    </>
+  )
+}
