+++ client/src/views/home.tsx
@@ -1,13 +1,13 @@
 'use client'
 
 import Link from "next/link"
-import Dev from "@/components/dev"
 import { comingSoon } from "@/constants/text"
 import Placeholders from "./placeholders"
 import useGaeldleStore from "@/stores/gaeldle-store"
 import { modesSlice } from "@/stores/modes-slice"
 import Modes from "@/components/modes"
 import Levels from "@/components/levels"
+import MainDev from "@/components/dev/main-dev"
 
 export default function Home() {
   const modesSliceState = useGaeldleStore() as modesSlice;
@@ -29,7 +29,7 @@ export default function Home() {
         <div className="mt-10 text-center">
           <p className="text-xl">{comingSoon}</p>
 
-          <Dev />
+          <MainDev />
         </div>
       </div>
     </main>
